+++ checks/container_rt.go
@@ -111,9 +111,9 @@ func fmtContainerStats(
 		sys2, sys1 := ctr.CPU.SystemUsage, lastCtr.CPU.SystemUsage
 		chunk = append(chunk, &model.ContainerStat{
 			Id:         ctr.ID,
-			UserPct:    calculateCtrPct(ctr.CPU.User, lastCtr.CPU.User, sys2, sys1, cpus, lastRun),
-			SystemPct:  calculateCtrPct(ctr.CPU.System, lastCtr.CPU.System, sys2, sys1, cpus, lastRun),
-			TotalPct:   calculateCtrPct(ctr.CPU.User+ctr.CPU.System, lastCtr.CPU.User+lastCtr.CPU.System, sys2, sys1, cpus, lastRun),
+			UserPct:    calculateCtrPct(uint64(ctr.CPU.User), uint64(lastCtr.CPU.User), sys2, sys1, cpus, lastRun),
+			SystemPct:  calculateCtrPct(uint64(ctr.CPU.System), uint64(lastCtr.CPU.System), sys2, sys1, cpus, lastRun),
+			TotalPct:   calculateCtrPct(uint64(ctr.CPU.User+ctr.CPU.System), uint64(lastCtr.CPU.User+lastCtr.CPU.System), sys2, sys1, cpus, lastRun),
 			CpuLimit:   float32(ctr.Limits.CPULimit),
 			MemRss:     ctr.Memory.RSS,
 			MemCache:   ctr.Memory.Cache,
